desafio/frontend-testing/requirements.txt
behave
PyHamcrest
selenium
pyautogui